css: max line length
Set the maximum line length to about 80 characters to make the wiki
easier to read on wide windows. The excessively long lines are hard
to read, and people have been putting line breaks in the pages
because of this.  Instead of sprinkling <br/> (or markdown lines
ending in two spaces) everywhere, use css to make the lines a
reasonable length.